Refactoring index.tsx. Changed Button component's way to get props in Button.tsx. Moved fullLocalStorage type to file where it is used first.
