+++ components/Button.tsx
@@ -1,33 +1,9 @@
-import { MouseEventHandler, ReactNode } from 'react';
+import { ButtonHTMLAttributes } from 'react';
 
 export function Button({
-  id,
-  handleClick,
-  type,
   children,
-  className,
-  onMouseOver,
-  onMouseOut,
-}: {
-  id?: string;
-  handleClick?: (event?: any) => void;
-  children?: ReactNode;
-  type: 'button' | 'submit' | 'reset' | undefined;
-  className?: string;
-  isDisabled?: boolean;
-  onMouseOver?: MouseEventHandler<HTMLButtonElement>;
-  onMouseOut?: MouseEventHandler<HTMLButtonElement>;
-}) {
-  return (
-    <button
-      id={id}
-      className={className}
-      type={type}
-      onClick={handleClick}
-      onMouseOver={onMouseOver}
-      onMouseOut={onMouseOut}
-    >
-      {children}
-    </button>
-  );
+  ...rest
+}: ButtonHTMLAttributes<HTMLButtonElement>) {
+  return <button {...rest}>{children}</button>;
 }
+
