+++ types/types.ts
@@ -1,7 +1,1 @@
-export type FullLocalStorage = {
-  name: string;
-  gift: string;
-  id?: string;
-  localStorageKeyID?: string;
-  createdDate: number;
-};
+
